CF/BF - Move CMS for RTC6705 into vtx_rtc6705.c for consistency with other VTX CMS menus.
