+++ src/main/cms/cms_menu_builtin.c
@@ -40,13 +40,13 @@
 
 #include "cms/cms_menu_imu.h"
 #include "cms/cms_menu_blackbox.h"
-#include "cms/cms_menu_vtx.h"
 #include "cms/cms_menu_osd.h"
 #include "cms/cms_menu_ledstrip.h"
 #include "cms/cms_menu_misc.h"
 
 // User supplied menus
 
+#include "io/vtx_rtc6705_cms.h"
 #include "io/vtx_smartaudio_cms.h"
 #include "io/vtx_tramp.h"
 
@@ -99,7 +99,7 @@ static OSD_Entry menuFeaturesEntries[] =
 #endif
 #if defined(VTX_CONTROL)
 #if defined(VTX_RTC6705)
-    {"VTX", OME_Submenu, cmsMenuChange, &cmsx_menuVtx, 0},
+    {"VTX", OME_Submenu, cmsMenuChange, &cmsx_menuVtxRTC6705, 0},
 #endif // VTX_RTC6705
 #if defined(VTX_SMARTAUDIO)
     {"VTX SA", OME_Submenu, cmsMenuChange, &cmsx_menuVtxSmartAudio, 0},
